+++ src/github.rs
@@ -1,16 +1,10 @@
 use crate::Project;
-use failure::{Error, Fail};
+use failure::Error;
 use log::debug;
 use log::trace;
 use reqwest::header;
 use serde_derive::Deserialize;
 use std::fmt::Display;
-
-#[derive(Debug, Fail)]
-pub(crate) enum GithubError {
-    #[fail(display = "Tag {} not found", tagname)]
-    TagNotPresent { tagname: String },
-}
 
 #[derive(Debug, Deserialize)]
 struct ReplyData {
@@ -60,9 +54,16 @@ impl Display for RepoStatus {
 }
 
 #[derive(Debug, Deserialize)]
+pub(crate) struct AssetJson {
+    pub(crate) id: u64,
+    pub(crate) name: String,
+}
+
+#[derive(Debug, Deserialize)]
 struct ReleaseJson {
     id: u64,
     tag_name: String,
+    assets: Vec<AssetJson>,
 }
 
 fn get_req_remaining(h: &header::HeaderMap) -> u32 {
@@ -109,33 +110,31 @@ pub(crate) fn get_status(prj: &Project, token: &str) -> Result<(RepoStatus, u32)
     Ok((rs, req_left))
 }
 
-fn _get_release_id_from_json(tag: &str, json: &str) -> Result<u64, Error> {
-    let rss: Vec<ReleaseJson> = serde_json::from_str(json)?;
-    if let Some(found) = rss.iter().find(|x| x.tag_name == tag) {
-        Ok(found.id)
-    } else {
-        Err(Error::from(GithubError::TagNotPresent {
-            tagname: tag.to_string(),
-        }))
-    }
+fn _get_release_id_from_json(_tag: &str, json: &str) -> Result<(u64, Vec<AssetJson>), Error> {
+    let rss: ReleaseJson = serde_json::from_str(json)?;
+    Ok((rss.id, rss.assets))
 }
 
-pub(crate) fn get_release_id(prj: &Project, tag: &str, token: &str) -> Result<(u64, u32), Error> {
+pub(crate) fn get_release_id(
+    prj: &Project,
+    tag: &str,
+    token: &str,
+) -> Result<(u64, Vec<AssetJson>, u32), Error> {
     let mut token_str = "token ".to_string();
     token_str.push_str(token);
     let mut h = header::HeaderMap::new();
     h.insert(reqwest::header::AUTHORIZATION, token_str.parse().unwrap());
     let url = format!(
-        "https://api.github.com/repos/{}/{}/releases",
-        prj.owner, prj.project
+        "https://api.github.com/repos/{}/{}/releases/tags/{}",
+        prj.owner, prj.project, tag
     );
     let client = reqwest::Client::builder().default_headers(h).build()?;
     let mut reply = client.get(&url).send()?;
     let req_left = get_req_remaining(&reply.headers());
     let json = reply.text()?;
     trace!("output is {:?}", json);
-    let release_id = _get_release_id_from_json(tag, &json)?;
-    Ok((release_id, req_left))
+    let (release_id, assets) = _get_release_id_from_json(tag, &json)?;
+    Ok((release_id, assets, req_left))
 }
 
 #[cfg(test)]
@@ -144,21 +143,48 @@ mod tests {
 
     #[test]
     fn test_get_release_id_from_json() {
-        let json = r#"[
+        let json = r#"
+        {
+            "url": "https://api.github.com/repos/pizzamig/ci-test/releases/15076202",
+            "id": 15076202,
+            "tag_name": "0.1.1",
+            "assets": [],
+            "author": {
+              "login": "pizzamig"
+            }
+        }
+        "#;
+        let (id, assets) = _get_release_id_from_json("0.1.1", json).unwrap();
+        assert!(id == 15_076_202);
+        assert!(assets.is_empty());
+    }
+
+    #[test]
+    fn test_get_release_id_from_json_assets() {
+        let json = r#"
         {
             "url": "https://api.github.com/repos/pizzamig/ci-test/releases/15076202",
             "id": 15076202,
             "tag_name": "0.1.1",
             "author": {
               "login": "pizzamig"
+            },
+            "assets": [
+            {
+                "id": 10739504,
+                "name": "FreeBSD-11.2-ci-test.tar.gz",
+                "download_count": 0
+            },
+            {
+                "id": 10739535,
+                "name": "FreeBSD-12.0-ci-test.tar.gz",
+                "download_count": 1
             }
-        },
-        {
-            "id": 15059566,
-            "tag_name": "0.1"
+            ]
         }
-        ]"#;
-        let id = _get_release_id_from_json("0.1.1", json).unwrap();
+        "#;
+        let (id, assets) = _get_release_id_from_json("0.1.1", json).unwrap();
         assert!(id == 15_076_202);
+        assert_eq!(assets.len(), 2);
     }
 }