+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - add a --tag-name option, to support automatic asset upload
 - add support to custom build template scripts
 - add support to the no_deploy clause in the yaml file
+- asset support: delete the asset, if already present, allowing "update"